% Generated by hand
\name{hwsd_download}
\alias{hwsd_download}
\title{Download HWSD v2.0 index grid}
\usage{
hwsd_download(ws_path = file.path(tempdir(), "ws_db"), verbose = FALSE)
}
\arguments{
\item{ws_path}{Path where the HWSD v2.0 grid will be stored.}
\item{verbose}{Show progress messages.}
}
\value{
The path used.
}
\description{
Downloads the HWSD v2.0 gridded index file for use with \code{hwsd_extract()}. The target directory is created if it does not exist.
}
\examples{
\dontrun{
# download once to a persistent path
hwsd_download(ws_path = "D:/data/HWSD2", verbose = TRUE)
}
}
